style: rename types to ad postfix "type | props"
