+++ apps/web/types/base-props.ts
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export type BaseProps = PropsWithChildren & {
+  className: string;
+};
