+++ apps/web/types/card-info-type.ts
