+++ apps/web/types/order-type.ts
@@ -1,4 +1,4 @@
-import { TProduct } from "./product";
+import { TProduct } from "./product-type";
 import { TUser } from "./user";
 
 export type TOrderStatus = 'processing' | 'shipped' | 'delivered' | 'cancelled' | 'returned';
@@ -8,6 +8,7 @@ export type TOrder = {
   user: TUser;
   products: TProduct[];
   total: number;
-  orderDate: string; // ISO 8601 string format
+  orderDate: Date;
+  updatedAt?: Date;
   status: TOrderStatus;
 };
